Update QuickSort.h
Corrected int to double changes

+++ QuickSort.h
@@ -2,8 +2,8 @@
 class QuickSort
 {
 public:
-	void swap(int* a, int* b);
-	void quickSort(double array[], double low, double high);
-	double partition(double array[], double low, double high);
+	void swap(double* a, double* b);
+	void quickSort(double array[], int low, int high);
+	int partition(double array[], int low, int high);
 };
